Put configuration error handling in correct place
